Add addToCartFromWishlist command to Merlins
